Broke annotate code into seperate files. Better error messages
