+++ reditools/annotate/__init__.py
@@ -1,3 +1,4 @@
 from .main import main
+from .rtannotater import RTAnnotater
 
-__all__ = ('main',)
+__all__ = ('main', 'RTAnnotater')
